backend/requirements.txt
asgi_lifespan
argon2_cffi
asyncio
bcrypt
black
bson
colorama
email-validator
fastapi
fastapi-mail
httpx
isort
loguru
motor
mypy
passlib
pre-commit
pyjwt
pymongo
python-dotenv
python-decouple
python-multipart
python-slugify
pytest-asyncio
pytest-cov
pytest-xdist
trio
uvicorn
